게시글 삭제 ux 개선
- confirm 추가
- refresh 추가

+++ vue-endgame/vue-til/src/components/posts/PostListItem.vue
@@ -26,7 +26,10 @@ export default {
   },
   methods: {
     async deleteItem() {
-      await deletePost(this.postItem._id);
+      if(confirm('진짜 삭제하실?')) {
+        await deletePost(this.postItem._id);
+        this.$emit('refresh');
+      }
     },
   }
 }
